+++ packages/core/test/client/client.test.ts
@@ -604,4 +604,68 @@ describe("createAuthClient", () => {
         })
         expect(response).toEqual({ success: false, tokens: null })
     })
+
+    test("getAccessToken with valid response", async () => {
+        const get = vi.fn()
+
+        get.mockResolvedValueOnce(createJSONResponse({ csrfToken: "csrf_token_1" }))
+
+        get.mockResolvedValueOnce(
+            createJSONResponse({
+                success: true,
+                tokens: {
+                    accessToken: "brand-new-refreshed-token",
+                },
+            })
+        )
+
+        createClientMock.mockReturnValue({
+            get,
+            post: vi.fn(),
+        })
+
+        const client = createAuthClient({ baseURL: "https://example.com" })
+        const response = await client.getAccessToken("github")
+
+        expect(get).toHaveBeenCalledTimes(2)
+        expect(get).toHaveBeenCalledWith("/csrfToken")
+        expect(get).toHaveBeenCalledWith("/providers/:oauth/tokens", {
+            params: { oauth: "github" },
+            headers: {
+                "X-CSRF-Token": "csrf_token_1",
+            },
+        })
+        expect(response).toEqual("brand-new-refreshed-token")
+    })
+
+    test("getAccessToken with invalid response", async () => {
+        const get = vi.fn()
+
+        get.mockResolvedValueOnce(createJSONResponse({ csrfToken: "csrf_token_1" }))
+
+        get.mockResolvedValueOnce(
+            createJSONResponse({
+                success: false,
+                accessToken: null,
+            })
+        )
+
+        createClientMock.mockReturnValue({
+            get,
+            post: vi.fn(),
+        })
+
+        const client = createAuthClient({ baseURL: "https://example.com" })
+        const response = await client.getAccessToken("github")
+
+        expect(get).toHaveBeenCalledTimes(2)
+        expect(get).toHaveBeenCalledWith("/csrfToken")
+        expect(get).toHaveBeenCalledWith("/providers/:oauth/tokens", {
+            params: { oauth: "github" },
+            headers: {
+                "X-CSRF-Token": "csrf_token_1",
+            },
+        })
+        expect(response).toEqual(null)
+    })
 })